+++ ckan/controllers/home.py
@@ -71,7 +71,7 @@ class HomeController(base.BaseController):
                 'groups': _('Groups'),
                 'tags': _('Tags'),
                 'res_format': _('Formats'),
-                'license': _('License'),
+                'license': _('Licenses'),
             }
 
             data_dict = {'sort': 'packages', 'all_fields': 1}