+++ ckan/logic/auth/delete.py
@@ -2,6 +2,7 @@ import ckan.logic as logic
 import ckan.new_authz as new_authz
 from ckan.logic.auth import get_package_object, get_group_object, get_related_object
 from ckan.logic.auth import get_resource_object
+import ckan.logic.auth.create as _auth_create
 from ckan.lib.base import _
 
 
@@ -130,19 +131,13 @@ def tag_delete(context, data_dict):
     # sysadmins only
     return {'success': False}
 
-def _group_or_org_member_delete(context, data_dict):
-    group = get_group_object(context, data_dict)
-    user = context['user']
-    authorized = new_authz.has_user_permission_for_group_or_org(
-        group.id, user, 'delete_member')
-    if not authorized:
-        return {'success': False, 'msg': _('User %s not authorized to delete organization %s members') % (user, group.id)}
-    else:
-        return {'success': True}
+def group_member_delete(context, data_dict):
+    ## just return true as logic runs through member_delete
     return {'success': True}
 
-def group_member_delete(context, data_dict):
-    return _group_or_org_member_delete(context, data_dict)
-
 def organization_member_delete(context, data_dict):
-    return _group_or_org_member_delete(context, data_dict)
+    ## just return true as logic runs through member_delete
+    return {'success': True}
+
+def member_delete(context, data_dict):
+    return _auth_create.member_create(context, data_dict)