+++ ckanext/example_iauthfunctions/tests/test_example_iauthfunctions.py
@@ -3,14 +3,88 @@
 '''
 import paste.fixture
 import pylons.test
+import pylons.config as config
+import webtest
 
 import ckan.model as model
 import ckan.tests as tests
-import ckan.plugins as plugins
-import ckan.plugins.toolkit as toolkit
+import ckan.plugins
+import ckan.new_tests.factories as factories
 
 
-class TestExampleIAuthFunctionsPlugin(object):
+class TestExampleIAuthFunctionsCustomConfigSetting(object):
+    '''Tests for the plugin_v5_custom_config_setting module.
+
+    '''
+    def _get_app(self, users_can_create_groups):
+
+        # Set the custom config option in pylons.config.
+        config['ckan.iauthfunctions.users_can_create_groups'] = (
+            users_can_create_groups)
+
+        # Return a test app with the custom config.
+        app = ckan.config.middleware.make_app(config['global_conf'], **config)
+        app = webtest.TestApp(app)
+
+        ckan.plugins.load('example_iauthfunctions_v5_custom_config_setting')
+
+        return app
+
+    def teardown(self):
+
+        # Remove the custom config option from pylons.config.
+        del config['ckan.iauthfunctions.users_can_create_groups']
+
+        # Delete any stuff that's been created in the db, so it doesn't
+        # interfere with the next test.
+        model.repo.rebuild_db()
+
+    @classmethod
+    def teardown_class(cls):
+        ckan.plugins.unload('example_iauthfunctions_v5_custom_config_setting')
+
+    def test_sysadmin_can_create_group_when_config_is_False(self):
+        app = self._get_app(users_can_create_groups=False)
+        sysadmin = factories.Sysadmin()
+
+        tests.call_action_api(app, 'group_create', name='test-group',
+                              apikey=sysadmin['apikey'])
+
+    def test_user_cannot_create_group_when_config_is_False(self):
+        app = self._get_app(users_can_create_groups=False)
+        user = factories.User()
+
+        tests.call_action_api(app, 'group_create', name='test-group',
+                              apikey=user['apikey'], status=403)
+
+    def test_visitor_cannot_create_group_when_config_is_False(self):
+        app = self._get_app(users_can_create_groups=False)
+
+        tests.call_action_api(app, 'group_create', name='test-group',
+                              status=403)
+
+    def test_sysadmin_can_create_group_when_config_is_True(self):
+        app = self._get_app(users_can_create_groups=True)
+        sysadmin = factories.Sysadmin()
+
+        tests.call_action_api(app, 'group_create', name='test-group',
+                              apikey=sysadmin['apikey'])
+
+    def test_user_can_create_group_when_config_is_True(self):
+        app = self._get_app(users_can_create_groups=True)
+        user = factories.User()
+
+        tests.call_action_api(app, 'group_create', name='test-group',
+                              apikey=user['apikey'])
+
+    def test_visitor_cannot_create_group_when_config_is_True(self):
+        app = self._get_app(users_can_create_groups=True)
+
+        tests.call_action_api(app, 'group_create', name='test-group',
+                              status=403)
+
+
+class TestExampleIAuthFunctionsPluginV4(object):
     '''Tests for the ckanext.example_iauthfunctions.plugin module.
 
     '''
@@ -24,17 +98,7 @@ class TestExampleIAuthFunctionsPlugin(object):
 
         # Test code should use CKAN's plugins.load() function to load plugins
         # to be tested.
-        plugins.load('example_iauthfunctions')
-
-    def setup(self):
-        '''Nose runs this method before each test method in our test class.'''
-
-        # Access CKAN's model directly (bad) to create a sysadmin user and save
-        # it against self for all test methods to access.
-        self.sysadmin = model.User(name='test_sysadmin', sysadmin=True)
-        model.Session.add(self.sysadmin)
-        model.Session.commit()
-        model.Session.remove()
+        ckan.plugins.load('example_iauthfunctions_v4')
 
     def teardown(self):
         '''Nose runs this method after each test method in our test class.'''
@@ -51,31 +115,25 @@ class TestExampleIAuthFunctionsPlugin(object):
         '''
         # We have to unload the plugin we loaded, so it doesn't affect any
         # tests that run after ours.
-        plugins.unload('example_iauthfunctions')
+        ckan.plugins.unload('example_iauthfunctions_v4')
 
     def _make_curators_group(self):
         '''This is a helper method for test methods to call when they want
         the 'curators' group to be created.
 
         '''
+        sysadmin = factories.Sysadmin()
+
         # Create a user who will *not* be a member of the curators group.
-        noncurator = tests.call_action_api(self.app, 'user_create',
-                                           apikey=self.sysadmin.apikey,
-                                           name='noncurator',
-                                           email='email',
-                                           password='password')
+        noncurator = factories.User()
 
         # Create a user who will be a member of the curators group.
-        curator = tests.call_action_api(self.app, 'user_create',
-                                        apikey=self.sysadmin.apikey,
-                                        name='curator',
-                                        email='email',
-                                        password='password')
+        curator = factories.User()
 
         # Create the curators group, with the 'curator' user as a member.
         users = [{'name': curator['name'], 'capacity': 'member'}]
         curators_group = tests.call_action_api(self.app, 'group_create',
-                                               apikey=self.sysadmin.apikey,
+                                               apikey=sysadmin['apikey'],
                                                name='curators',
                                                users=users)
 
@@ -85,12 +143,14 @@ class TestExampleIAuthFunctionsPlugin(object):
         '''Test that group_create doesn't crash when there's no curators group.
 
         '''
+        sysadmin = factories.Sysadmin()
+
         # Make sure there's no curators group.
         assert 'curators' not in tests.call_action_api(self.app, 'group_list')
 
         # Make our sysadmin user create a group. CKAN should not crash.
         tests.call_action_api(self.app, 'group_create', name='test-group',
-                              apikey=self.sysadmin.apikey)
+                              apikey=sysadmin['apikey'])
 
     def test_group_create_with_visitor(self):
         '''A visitor (not logged in) should not be able to create a group.
@@ -129,18 +189,18 @@ class TestExampleIAuthFunctionsPlugin(object):
         assert result['name'] == name
 
 
-class TestExampleIAuthFunctionsPluginV3(TestExampleIAuthFunctionsPlugin):
+class TestExampleIAuthFunctionsPluginV3(TestExampleIAuthFunctionsPluginV4):
     '''Tests for the ckanext.example_iauthfunctions.plugin_v3 module.
 
     '''
     @classmethod
     def setup_class(cls):
         cls.app = paste.fixture.TestApp(pylons.test.pylonsapp)
-        plugins.load('example_iauthfunctions_v3')
+        ckan.plugins.load('example_iauthfunctions_v3')
 
     @classmethod
     def teardown_class(cls):
-        plugins.unload('example_iauthfunctions_v3')
+        ckan.plugins.unload('example_iauthfunctions_v3')
 
     def test_group_create_with_no_curators_group(self):
         '''Test that group_create returns a 404 when there's no curators group.
@@ -150,11 +210,9 @@ class TestExampleIAuthFunctionsPluginV3(TestExampleIAuthFunctionsPlugin):
 
         '''
         assert 'curators' not in tests.call_action_api(self.app, 'group_list')
-        user = tests.call_action_api(self.app, 'user_create',
-                                     apikey=self.sysadmin.apikey,
-                                     name='test-user',
-                                     email='email',
-                                     password='password')
+
+        user = factories.User()
+
         response = tests.call_action_api(self.app, 'group_create',
                                          name='test_group',
                                          apikey=user['apikey'], status=404)
@@ -175,18 +233,18 @@ class TestExampleIAuthFunctionsPluginV3(TestExampleIAuthFunctionsPlugin):
         assert response['__type'] == 'Authorization Error'
 
 
-class TestExampleIAuthFunctionsPluginV2(TestExampleIAuthFunctionsPlugin):
+class TestExampleIAuthFunctionsPluginV2(TestExampleIAuthFunctionsPluginV4):
     '''Tests for the ckanext.example_iauthfunctions.plugin_v2 module.
 
     '''
     @classmethod
     def setup_class(cls):
         cls.app = paste.fixture.TestApp(pylons.test.pylonsapp)
-        plugins.load('example_iauthfunctions_v2')
+        ckan.plugins.load('example_iauthfunctions_v2')
 
     @classmethod
     def teardown_class(cls):
-        plugins.unload('example_iauthfunctions_v2')
+        ckan.plugins.unload('example_iauthfunctions_v2')
 
     def test_group_create_with_curator(self):
         '''Test that a curator can*not* create a group.