+++ ckanext/example_theme/v02_empty_template/plugin.py
@@ -16,4 +16,6 @@ class ExampleThemePlugin(plugins.SingletonPlugin):
 
         # Add this plugin's templates dir to CKAN's extra_template_paths, so
         # that CKAN will use this plugin's custom templates.
+        # 'templates' is the path to the templates dir, relative to this
+        # plugin.py file.
         toolkit.add_template_directory(config, 'templates')